v1.1.3 - fix locked roles
When a user edited the roles through the Customizations UI, it was setting all Teams roles to is_editable:0, making them all locked. This was because the plugin wasn't overriding the roles in the filter to get roles.

Removed the checks to see if the roles already exist so that they get overridden with the expected value (undefined) for is_editable. Customized permissions still get set in a lower priority filter.